Move stageSizeToTransform into screen-utils
